fix datatype for value parameter

+++ CurrencyConverter.java
@@ -1,5 +1,5 @@
 
 
 interface CurrencyConverter{
-   Currency convert(Currency source,Currency destination, long value);
+   Currency convert(Currency source,Currency destination, double value);
 }
